switch back to wsgiref as a stopgap
Somehow I didn't realize that gunicorn always forks. We were sort of
getting away with this prior to tx_model since the (now defunct)
RouterTransaction output stack was spawned from the handler. With the
tx_model changes, it's spawned by dequeue in a thread from the startup
code. So we're in a state where the handler is running in the gunicorn
worker and the backend code is running in the parent. Incredibly,
since they communicate through the database, it kinda sorta works
modulo some polling delays. I'm actually stunned it works at all! I
would expect it to immediately crash or deadlock forking, etc.

It should be simple to reorganize the startup logic to run basically
router_service.Service.main from the gunicorn post_worker_init() hook.