replace BlobContent with a single BLOB field in Blob
sqlite3 (and at least pgsql) claim to support a max row of something
like 2GiB.  I did some manual testing that sqlite3 doesn't appear to
have to read the whole row into memory to do a concat or substr. So
this should be scalable enough for now and simpler than striping the
data by hand.

It seems like the result of concat/substr on BLOB is TEXT somewhere
between sqlite3 or the python3 wrappers so we end up needing
CAST(... AS BLOB) to work correctly. Added a test with random binary
data to convince myself that this is a noop and not round-tripping
through unicode at some point.