storage: move update of tx attempt count from start_attempt to finalize
The way it is, a tx won't be recovered if it crashed during its final
attempt so upcoming bounce logic wouldn't get a chance to run. I think
at the time, the idea was not to recover oneshot transactions in case
the client went away so we would start it upstream again but I think
that is a relatively minor issue. Updated TODO with a possible
strategy to fix that.

Fix bug that Storage._write() would set max_attempts=1 if not set in
tx delta! Update to set to set to any value in delta including 0.