Replace custom implementation of tx delta/merge in RestEndpoint with more uniform treatment in TransactionMetadata
This finally reached the breaking point with the rest representation
POST {
  'mail_from': { ... }
}
->
{
  'mail_response': {}
}
to denote accepted/inflight

change to
POST {
  'mail_from': { ... }
}
->
{
  'mail_from': {}
}

TxField emit/accept were not clearly thought through, replace with a
single "validity." Note that REST_CREATE/UPDATE are for the *request*,
for now we think the reply should be identical to REST_READ which is
how it's actually being used. If that turns out to be insufficient,
there can be separate REST_CREATE_REQ/RESP etc.