The current Filter delta representation has proven to be unwieldy and result in complicated and error-prone diff/merge any time the filter needs to send a modified transaction upstream.
Start building a replacement SyncFilter that takes both the full
vector and the delta and returns the delta. Port ReceivedHeaderFilter
as a test case. Implement Filter->SyncFilter and vice versa adapters
for chaining. New FakeSyncFilter is more of a mock that takes a
callable to inspect the contents of the downstream call and return the
upstream result.

The ad hoc code in ReceivedHeaderFilter to merge upstream responses
that we replaced here is typical of output chain filters. This is more
compact/complete/correct.

The new ReceivedHeaderFilter.on_update() was a little subtle. We
suspect we will want that exact logic in other situations and may end
up writing a PayloadFilter that takes a callable to either produce a
new body or return an error.