save offset of list fields in deltas, merge fail if the offset doesn't match   the bug of repeatedly applying the same delta has recurred repeatedly!
this also ends up in json for REST_UPDATE with
rcpt_to_list_offset: 2
etc.

This was already ~poor man's json-patch and while this brings us a
little closer. I think ours is still such a tiny subset of the
functionality that switching to that wouldn't be a huge win here.