fix message builder 1: regression from rest handler refactor: need to call MessageBuilder.get_blobs() in StorageWriterFilter   add to router_service_test
2: TransactionBlobRefs table was never read except for reachability
for gc. As such, there is nothing to ensure that a blob id in a
message builder spec is actually referenced to that tx via
blobrefs. Pass the tx id to BlobReader.load() to enforce this.

In the first cut of this patch, we weren't adding the blobrefs in
MessageBuilderFilter but it would work anyway at least initially but
there's nothing to prevent the blob from getting gc'd in the meantime.

3: storage don't early return in _write_blob if no ids so
input_done gets set when adding message_builder with no blobs

4: storage._load_db() save db tx id int in tx

5: move blob<->uri to common location in rest_service

6: add etags to send_message.py