RestEndpointAdapter -> StorageWriterFilter -> Storage refactoring/refinement
Previously, after writing a blob via AsyncFilter.get_blob_writer(),
you still had to write the blob to the tx via
TxCursor.write_envelope(
  TransactionCursor(body=<blob_rest_id>),
  reuse_blob_rest_ids=[<blob_rest_id>])
and this provided the "wakeup" to write input_done, etc.

  Aside: this is not necessary for message_builder which always
  requires finalized blobs

1: Storage._write_blob() read tx blobrefs first, skip blobs that are
already ref'd in the tx.

2: TxCursor._write() take new parameter require_finalized_blobs. This
is always true from TxCursor.write_envelope() but is relaxed from
Storage.create_blob().

3: BlobWriter take new params update_tx and finalize_tx. If set,
update_tx does an empty TxCursor._write() after each blob write to
ping last_update/version. If finalize_tx is set (when writing
tx_body), this will also set input_done when the blob is finalized.

The net of this is
1: blobs are ref'd in the transaction when they are created
2: the transaction last_update/version updates every time blobs are written
  useful to know the downstream is making progress for timeouts
3: the wakeups no longer need to be provided further up the stack
  i.e. AsyncFilter.update() calls removed from RestEndpointAdapter, etc.
  the flip side of that is the WritableBlob impl has to do that instead
4: the internal submission flows (Exploder, Notifications) that write
body_blob don't change for the moment but we should revisit

I think the "v2/orthogonal blobs" approach was trying a little too
hard to treat the body as a special case of the message_builder which
was great for code commonality but the rest presentation was perhaps a
little weird. I'm hoping this will be a good tradeoff between adding a
little more special-case code for the body vs message builder blobs in
order to provide a first-class experience in the common case.

TODO
sort out the exact rest presentation for some of the reuse cases
add router_service_test
- body reuse
- message builder blob reuse