Hal fixing a light bulb
1: manual testing revealed that retry timeout wasn't emitting a
bounce. This turned out to be a bug in the notification_workflow
changes that OutputHandler._next_attempt_time was returning
final_attempt_reason "oneshot" if retries weren't enabled at the
time. The net of this was that this didn't get into the path in
TxCursor._write() to set no_final_notification, etc. Ditto for the
no_final_notification case in OutputHandler.handle()
TODO finer-grained test of this in OutputHandlerTest

2: router_service_test had this exact scenario (retry tempfail -> bounce)
but wasn't ticking this bug due to a combination of
- sqlite limiting concurrency
- test dequeueing upstream tx after downstream has already timed out, etc
such that we don't exercise the typical case where the exploder and
upstream OH ping-pong on the db tx

So switch router_service_test to postgres and touch up the other
tests. This uncovered more issues.

3: due to misunderstanding the level of isolation provided by the db
(READ COMMITTED, not REPEATABLE READ), it was possible for
Storage.create_blob() to see a version conflict. In particular, the
upstream OH was writing an upstream envelope response concurrent with
the downstream rest handler writing an inline body. Retry these in
StorageWriterFilter.  Ditto for Storage.load_one(). Switching the db
engine to the higher isolation mode just causes db_tx.execute() to throw
concurrency exceptions from the underlying dbapi, in this case psycopg.

4: test_retry -> test_rest_smoke
This test was previously effectively doing gw/interactive smtp
protocol without the exploder which is not how it's actually
used. Shave this down to a minimal smoke test now that we have better
coverage of retries with the other exploder/notification tests.

5: update assertions in TxCursor._write_blob() to make it more obvious
that it actually works