start reworking top-level router workflow rest and exploder upstream tx creation paths make an upcall to the top-level router to get the StorageWriterFilter. This now starts the output handler synchronously or fails.
It is the exploder that emits downstream 250 store&forward
responses so if it can't start, the downstream will just see
timeouts. It is more arguable for exploder upstream for msa but IMO
you should be able to operate with plenty of headroom for submission
and use front proxy/envoy/etc to fail over on the 500 if it's
overloaded. mx should be prepared for next-hop busy errors.

TODO
- ensure Storage.load_one() cannot race with this and dequeue
  freshly-created tx
- dismantle Storage.wait_created() machinery