schedule remaining handler methods
fix brown bag bug in Executor not propagating rv of lambda,
unbelievable I didn't encounter this until today!

This should be working on both flask and fastapi at this point. I will
probably keep both around for awhile until I get some more experience
with fastapi. It's possible the flask implementation doesn't work well
under load as-is. I'm not sure which executor hypercorn runs wsgi on,
etc.

The next step here could be to decouple hanging GET waiting in
RestHandler so e.g. an update schedules the result separately that can
fall back to polling if too overloaded to add more waiters. Or go
direct to async waiting for that...