re-integrate IdVersionMap with Storage   this came out a little differently than I expected when I started   refactoring this
For the rest read/if-none-match path: allow creating a
TransactionCursor by id without loading. This samples the current
version in the cache in cursor.version

Somehow this exposed what seems to be a latent race that e.g. an write
could update the version cache in between a load reading the db and
updating the cache. Just raise VersionConflictException for now and
retry in the calling code.

Fix horrifying bugs in storage where if neither db_id nor rest_id was
specified, it would do the query anyway without the WHERE constraint and
return an ~arbitrary row \o/

Tweak nospin logic in RestEndpoint._get() not to wait if we got a
different etag back from the server no matter how short the delta
is. Ditto for examples/cli/send_message.py