remove db tx body_blob_id/body_rest_id cols, these are vestigal, we always get this from blobrefs now
If this wasn't an effect of the previous patch, it definitely is the
case now: it is no longer possible to reuse a tx body as a message
builder blob or vice versa. Again, this isn't a filesystem, this is
only to say "I want to send 10 transactions with the same body", not
arbitrary many-to-many relationships.

leave a minor kludge: TxCursor.load populates tx.body with a
placeholder string which downstream rest read path converts to rest
placeholder {}

don't allow blob creates external to storage in __internal.* namespace