fix "downstream delta" bugs in a bunch of filters, discovered yet another wrinkle which is that you often need if bool(delta):   upstream.on_update(...) to not send empty deltas which is invalid
change rest endpoint to look at upstream_tx instead of caller tx for
'all rcpts failed' check which was the symptom here. Though in the
presence of the filter bugs, this shifted the symptom from requests
failing with "failed precondition: all rcpts failed" from RestEndpoint
to an delta/merge assertion somewhere else in the stack. That's at
least more conspicuous so maybe that's an improvement?

not happy with how error prone this seems to be, need some kind of
"filter compliance" testing, debug mode that wraps every filter in a
checker that makes sure they don't drop fields like these bugs