fix a cluster of things that bubbled up from last deploy bad config was missing dns_resolution so we were sending reqs to gateway with 'remote_host': ['gmail.com', 25]
smtplib resolves that to an A but that's the vip for the web ui which
(weirdly?) seems to silently discard port 25 instead of RST so the
connect would hang forever

1: always set a timeout in smtplib
2: gateway default yaml watchdog timeout -> 1h, smtp timeout 55 min
3: SyncFilterAdapter had diverged from AsyncFilter semantics:
- update(), get() do not do upstream inflight waiting, drop timeout
param which is not in abc
4: SmtpEndpoint reject remote host that is not a valid ascii IP
5: yet another VersionConflictException on cancellation, this one in
StorageWriterFilter._update()

6: SyncFilterAdapter concurrency control fixes: there was a race where
a get() could see a torn write of tx/version which would cause the
downstream to hang since it had the new etag but the old tx that
showed the req still inflight. Update sampled_version in
get()/update(), version() assert that it's non-None.

Remove _version, was seemingly a bug that the increment in update()
wasn't being propagated?!

Actually, this api makes no sense in the presence of concurrent rest
reqs to SyncFilterAdapter. update/get need to return version directly!