Executor spawn threading.Thread directly instead of ThreadPoolExecutor
ThreadPoolExecutor always runs the worker threads as daemon=False and
even installs an atexit handler to join all of the workers. So afaict
it is impossible to exit the process if a thread has hung forever due
to a bug; you have to have an separate process sidecar/supervisor
thing to SIGKILL it. But we actually want to die on a watchdog
timeout.

https://stackoverflow.com/questions/49992329/the-workers-in-threadpoolexecutor-is-not-really-daemon

https://github.com/python/cpython/issues/83993

We can revisit at such time we become cpu bound on spawning new
threads every time vs reusing!

Adjacent changes:
- Executor.submit() do not check_watchdog()
- router and gateway install heartbeat functions with hypercorn_main
  to do that instead