disallow multi-rcpt lmtp
I don't have any firsthand experience with LMTP but according to rfc2033
it returns a separate DATA response for each successful RCPT. However
smtplib/aiosmtpd don't appear to support this. It may be that the only
place this is ever used is for an MTA to deliver to an application
where it's already been exploded by the MTA so this may not be a
limitation in practice.