Storage reject notifications without retries enabled
Messing around with the sending examples, I created a transaction with
notifications enabled and retries=None and hilarity ensued because of
a surprising interaction with the oneshot/no_final_notification
workflows for the Exploder. Basically the absence of retry causes
start_attempt to preload final_attempt_reason with "oneshot" but then
the presence of notification causes the write at the end of the
OutputHandler to clear it so it will retry forever or at least until
it succeeds. Also the absence of retry= causes early-return from the
next_attempt_time logic so the retries will have no delay.

My conclusion is that it doesn't really make sense to request
notifications for an ephemeral transaction e.g. do you expect to get a
notification if the router crashes in the middle of the attempt? Prior
to 6c7b8738 we had a more explicit notion of durable vs ephemeral
transactions like ephemeral might be backed by tmpfs, etc. Instead
request retry={"max_attempts": 1} which means "get to the end of the
OutputHandler once".

Longer term:
- now that we're on fastapi, look into using pydantic for request validation
- these special storage tx workflows/lifecycles are difficult to
  reason about, think about how this could be simplified