better downstream timeout behavior for submission
If the upstream transaction fastfails, OutputHandler will keep waiting
for the downstream body before returning which isn't great.

1: remove input_done condition from have_body so we will early return
as soon as we get any of the body if none of the recipients have
succeeded
2: don't set final_attempt_reason = 'downstream timeout' in this case
if retries are enabled

This way, at such time as the client finishes uploading the body, the
tx will become input_done=true and thus recoverable.