db timekeeping improvements
always store timestamps in db as integer unix secs
add Blob.creation, TransactionAttempts.creation and last_update
add NOT NULL constraints

Add Storage._current_timestamp_epoch() which returns a sa subquery that
returns unix secs from sql current timestamp using engine-specific
sql. Update all call sites to use that instead of local time at the
application instance.

There doesn't seem to be a portable solution to this; this is the best
I can come up with that minimizes engine-specific code. In particular sqlite's
behavior can be pretty counterintuitive e.g.
> select current_timestamp;
2024-12-11 01:20:50
> typeof(current_timestamp)
text
but if you try to take a difference of 2 timestamps, it seems to do
the equivalent of
> CAST(current_timestamp AS INTEGER)
2024
which appears to be lexing out up to the first non-digit character
which from the above example is the year so it does the arithmetic on the years
whereas sql92 gives something like
TIMESTAMP - TIMESTAMP -> INTERVAL

Possible followup: update these timestamp columns with triggers so
it's automatic; we found several updates that were missing it. However
you still need the subquery trick here in any query where you want to
compute the time-since-last-update of something in the db.

gc ttl -> timedelta