router service test passing!
1: async filter wrapper don't set data precondition failure unless all
rcpt responses are present

2: storage.TransactionCursor.create() and write_envelope() had a
subtle behavior difference: create() used body_blob_uri() to add
internal body rest id and write_envelope() didn't. The net of the
exploder changes is that it previously always wrote the body in a
seprate update even if it got it all at once and now it writes it all
at once if possible. This moved the body blob write from the update
path to the create path.

In conjunction with this, StorageWriterFilter._get_body_blob_uri()
regenerate the BlobUri from a BlobCursor to drop the internal rest id
rather than passing through verbatim. Otherwise it hits an assert in
body_blob_uri() that the input rest id isn't internal.