all tests passing!
Managed to get the version conflict retry loop in async filter wrapper
to loop excessively. It seems like the net of these changes has caused
a behavior change that StorageWriterFilter._maybe_write_body_blob() is
now slowpath copying the bytes out of a BlobCursor instead of
cloning. In this case, the BlobCursor append pings the tx which causes
the downstream cursor.write_envelope() in update() to conflict
100%. This needs to be fully understood.