refactor router_service.Service.wait_shutdown()
Ping watchdog every 30s if waiting longer than that. Return _shutdown
so we don't access it without lock which is at least bad hygiene if
not an actual source of bugs.

daemon_executor has hard-coded watchdog timeout of 5m. I was using the
same gc interval on my testbed and actually saw watchdog timeouts in
gc.