fill_inflight_responses() instead of setting individual fields
The immediate issue this fixes is remote_host was only setting
mail_response after dns err but this is upstream from exploder so it
will always have rcpt at that point. Then rcpt gets 'internal error:
OutputHandler failed to populate response' which is what downstream
sees.

relay_auth is chained downstream of exploder so	this is	moot in	the
absence of pipelining but fix for completeness.

discussion: in some ways, it would be better to return 503 5.5.1 bad
sequence of commands/failed precondition for the responses after the
one that failed? But it's a little tricky to reason about whether the
client will see the real error around Exploder/AsyncFilterWrapper
i.e. a mail error upstream of the exploder probably won't get back to
the client?

Add note to AsyncFilterWrapper._set_precondition_resp() that may be
dead code per above