body/message_builder output workflow improvements
1: bug: storage skip _maybe_write_blob() if body not in delta, was
spuriously setting input_done
2: latent bug/quirk: storage _load_blobs() ORDER BY id
  the new body delta stuff in this patch requires this
3: OutputHandler do not clear body if not finalized
4: TransactionMetadata.delta() emit body if it grew but detect/fail if
it changed in some other unexpected way
5: message_builder_filter noop if message builder spec not finalized

The net of this is to give message_builder_filter a chance to reject
the json on the first update call before we possibly hang on the
upstream so it can report downstream to the client ~immediately.
TODO: implement this.