smtp client body streaming router -> gw -> wire
Add trivial flow control to SyncFilterAdapter: don't allow queueing
multiple chunks from downstream.

Add a copy of our fork of cpython/Lib/smtplib that supports payload
chunking and bdat. This should be identical to the one at
http://github.com/jsbucy/cpython/tree/smtplib_bdat and I plan to
maintain it there. In particular, I'm not shipping the smtplib tests
here because they depend on some test support stuff from the cpython
tree that's kind of a pain to deal with the toolchain in bazel, etc.

I don't have a strong opinion as to when would be a good time to try to
merge this upstream. I expect the rate of change upstream to be very
low (~1/year) so it's probably fine to track upstream manually for
the time being.

Refactor smtp client factory machinery to be able to take the smtplib
module as a parameter; use ours or the system one per yaml
smtp_output.use_system_smtplib=False

drop use_fastapi param dimension from gateway_test, this flag has been
ignored since f296b210