recipient router filter noop after previous failure
What seems to be happening on the testbed is that a spammer tries to
relay, recipient router filter sends them a 550 and then they go out
to lunch. 30s later, the output handler sends a noop/heartbeat with an
empty delta which router filter should early return. Instead it tries
to reapply the previous dest_delta which fails but is ignored and then
calls the upstream and hilarity ensues.

Probably TransactionMetadata.merge() failures should raise ValueError
since it's almost always a bug and we've been fairly inconsistent with
assert tx.merge_from(delta) is not None
which is repetitive anyway.