add TransactionMetadata.check_preconditions() for FilterChain
Invoke this after each filter invocation in FilterChain and stop if
the tx cannot make forward progress. As a result, filter
implementations no longer need to populate all responses.

Add asserts that responses aren't going upstream; I believe this is
always a bug.

check_preconditions() is a streamlined version of
fill_inflight_responses() and I don't want to yak-shave all of its
callers right now. But do the first bit of that in RestEndpoint. Many
uses of fill_inflight_responses() may be a little sloppy. The example
in Exploder was sort-of wrong even if in practice the mail err
couldn't end up in the downstream tx.