FilterChain break on not req_inflight()
otherwise we keep going after e.g. rcpt router filter sets an error
rcpt_response and the next iteration asserts on the downstream
response

This is a stopgap to fix an exception on my testbed.  This forces
other tx fields to ride along with a mail/rcpt/data request field. I
now believe the current logic in recipient router is incorrect but
that fix is a little more involved:
- recipient router filter should clear the recipients that failed (like
  RejectFirstRcpt filter in the test) and set the errors in the
  downstream path
- chain recipient router filter downstream of exploder to skip the
  upstream tx if the address doesn't route

add RejectFirstRcpt in filter_chain_test, an example of a filter that rejects
individual rcpts that could be chained downstream of exploder