support chaining rcpt router downstream of exploder
This allows rejecting invalid rcpts in the downstream flow before the
Exploder starts the upstream flow.

1: noop logic to stick a bool in the rcpt_to Mailbox instead of
top-level tx fields.
2: add bool 'dry_run' to recipient router filter yaml. This causes the
filter not to populate routing result tx.rest_endpoint/resolution
fields for use in the downstream side where we just want it to reject
invalid rcpts.

At first glance, it's appealing to save the routing results into the
rcpt_to Mailbox here. However our one-off IDL doesn't track field
validity across the whole tree; only the top-level fields in the
tx. So we can't control the propagation of individual fields within
Mailbox, we don't want to send them to other rest endpoints, etc. So
as it is, the routing result fields don't get written to the db and we
do it again in the upstream flow. That probably isn't a big
deal. Worst case you add a cache for the routing results. And you may
not want to reuse it across retries in case it gets stale. This in
isolation doesn't seem worth a big upgrade to the IDL.

Longer-term TODO: eliminate the reptition of the routing policy yaml.