start porting to rules_mypy
https://github.com/theoremlp/rules_mypy

Google announced this week that pytype is EOL
https://github.com/google/pytype/blob/main/README.md

There was never open-source bazel integration for pytype and indeed
this was difficult because of the (AFAICT) lack of a way to run it
incrementally; at least mine scaled quadratically
https://github.com/google/pytype/issues/1864
This became so slow even on this ~20kloc codebase that I would only
run it once as the last step before merge which reduces its utility
significantly.

Of course mypy is flagging some different stuff now that we will need
to go fix.