WIP about 75% of targets passing mypy, the rest disabled
By far the biggest difference is that pytype sort of fails open on
Optional/None and mypy always flags it. This results in adding a ton of
assert x is not None
e.g. downstream_tx in every single filter implementation which I'm not
exactly happy about but it is what it is

rules_mypy checks the tests to which I hadn't been doing before and
much to my surprise many of them were passing to begin with