hypercorn -> uvicorn
...which is under more active development, by the same team as
starlette/fastcgi and httpx

For the record, hypercorn has a quirk in the way it uses python
logging... it disables "propagate" on its loggers so you need to
configure them explicitly in logging.dictConfig or else they go
nowhere including handler exceptions. I fixed this for the rest
receiver in my testbed but never fixed it for router/gw.

configure uvicorn with an empty logging dict so it just logs to the
debug log the same as everything else