VersionCache only handoff cursor from update to already-waiting readers
Storage._load_db() always populate tx.attempt_count; attempt is
initially created with NULL TransactionAttempts.responses.

All tests passing but this is now too conservative and as eliminated a
lot of the improvements to the microbenchmark we had gotten up to this
point. We will probably put the caching behavior back at least for the
RestHandler reads under the constraint that the tx is currently leased
in this process.