WIP improved RestHandler.get_tx_async() workflow
put back cache semantics in VersionCache but only if it's leased
locally

Add new tx_cursor.check() which does a (significantly) cheaper query
to return only the tx version and lessee. For the cluster case, we
previously thought: if you're going to have to read to get the lessee,
you might as well just service the read. I've changed my mind
for now: do the cheaper read and redirect to the current lessee if
any.

Add some comments around caching http redirects. This is not really in
scope for router/gw but advantageous for lro clients. In particular:
- router should return 308 instead of 307
- client always starts with uri they got back from the original tx
  creation POST 201 but should use httpx.Client with redirect caching
  for each "session" of waiting on the LRO. This way there will be at
  most one cache miss per session. To scale to a large number of
  replicas, you might want to distribute the cache of which tx are leased in
  which replicas in something like memcache/redis.