add ttl caching to VersionCache
As it is, typically the last OH write sets data_response and finalize
attempt/final_attempt reason. OH then exits and unrefs IdVersion so it
falls out of IdVersionMap weak value dict. Even if a read was already
waiting, we won't reuse the cursor/tx and have to read from the db
again.

Add a set to IdVersionMap that holds refs to all of the IdVersions
that haven't expired ttl. IdVersion now reuses cursor if it's still leased
or it has final_attempt_reason. IdVersionMap.get() GCs entries that haven't
been updated within the ttl.

        1    0.000    0.000    0.000    0.000 2025-09-24 13:40:37,958 [138223072325760] router_service_test.py:1125 done <bound method RouterServiceTest._rest_smoke_micro of <__main__.RouterServiceTestFastApi testMethod=disabled_test_micro>> 5.114386 19.552690
90 -> 51!

        6 2025-09-24 13:40:48,266 [138223072325760] router_service_test.py:1125 done <bound method RouterServiceTest._exploder_micro of <__main__.RouterServiceTestFastApi testMethod=disabled_test_micro>> 10.149876 9.852337
200 -> 101!

Remove TransactionMetadata.tx_db_id. This was a hack related to message
builder that no longer uses it e6eae358