remove hack of stuffing version in tx
... in particular in StorageWriterFilter

This breaks the model that the tx is append only and begets many more
hacks where it has to be deleted before delta/merge to avoid a
conflict.

Astute readers may point out that the responses can create the same
problem however one must track the tx for much longer (across a retry)
to observe that changing.

We've talked before about having a concept of "live status fields" but
this would need some first-class machinery in the TransactionMetadata
to support properly without breeding hacks like this.

Rework TransactionCursor logic around id_version propagation on copy;
this should not end up on a copy going into VersionCache but should be
copied otherwise.