WIP step 1 of making rest mutations fully sticky: blobs
rest presentation of blobs (body/message_builder) switch from
placeholder {} to full status struct
{ length: 123, content_length: 123, finalized: true,
  uri: 'http://replica1/tx/123/body'}
etc.

This saves clients from having to manipulate these blob uris and just
send to that uri if !finalized. This also makes it easier to debug e.g that
your message isn't sending because one of the blobs isn't finalized.

On create, blob uris are qualified to the replica.

If the tx fastfailed (say on rcpt) and goes unleased, that uri becomes
qualified to the cluster on a subsequent GET.

TODO
- all req types 308 to current lessee if mismatch
  but continue to service in place if unleased
- it should not be possible for the tx to become leased again
(!input_done) but mutations/blob append with session mismatch should
fail in storage

This should eliminate the cross-replica ping to invalidate version
cache/trigger OH wakeup after a mutation landed in the wrong place.