drop MessageBuilderSpec.check_ids()
add type assertion that mypy wasn't flagging (to my surprise)

This been superseded by parse_blob_specs() to walk the mime part tree
in the message builder spec json and validate the parts. The storage
load path does this again as part of tx.from_json().

It is possible for a bug in this to result in a dangling blob id in
the mime part tree. This would then trip the check we're updating in
MessageBuilder._add_part() and throw. Add TODO to
MessageBuilderSpec.set_blobs() to do some more validation of this
contingent on having a way to quarantine invalid transactions in
storage.