fix hack in blob_spec_from_json(), bug in parse_blob_uri()
blob_spec_from_json() use parse_blob_uri() instead of
constructing BlobUri with fake/placeholder values

add BlobUri.delta() this should only be invoked when RestEndpoint
takes deltas between successive versions of the upstream

parse_blob_uri() wasn't saving parsed_uri in the tx_body case

Once we started doing that, RouterServiceTest.test_reuse_body started
failing with a bad delta in RestEndpoint uploading the 2nd tx to reuse
the body from the first. Apparently fixing parsed_uri caused the rest
handler to start returning it on reuse when it didn't before. This is
scoped to the current transaction rather than the one it reused it
from hence the delta. In examples/send_message, the condition for
whether it was reused successfully is simply
blob_json['finalized'] is True

port RouterServiceTest.test_reuse_body to examples/send_message, this
was using RestEndpoint to emit blob reuse which it doesn't normally do
which was tripping over the url change above