refactor BlobCursor._append_data()
fix TODO: the WHERE conditions on the UPDATE are sufficient input
validation and we don't need the SELECT in the common case where
the offset is correct

However we can easily handle the situation where there was a partial
update transferring a blob. So do the SELECT if the UPDATE failed
since we need the result to figure out whether we're in the resync
case or the invalid-offset case.

This way the client just needs to retry the PUT after an http error
and it should Just Work.