add filter.TxField.is_dict and TransactionMetadata.filter_output
is_dict TxFields are diff'd/merge'd

tx.filter_output (can we think of a better name? features?) is the new
home for metadata/derived data that is not "intrinsic" to handling the
transaction. The key is the fully qualified filter name
e.g. koukan.remote_host_filter.RemoteHostFilter. The value is a
filter-specific type. My hunch is that this is preferable to
dynamically extending TransactionMetadata and losing typing on the
core/intrinsic fields there.

drive-by cleanups
- tx.copy() use copy_valid_from() instead of copy.copy()
- Switch some checks from the corresponding is_list case in delta/merge
  to use TxField rather than isinstance(x, list)
- invert merge/merge_from: merge_from (real impl) throws and merge
  (wrapper) catches and returns None, in most cases the error is not
  expected i.e. terminate the handler and log a stack trace from the
  site of the error
- remove maybe_merge_from() which appears to be dead code